+++ builders/CacheBuilder.d.ts
@@ -4,10 +4,12 @@ export default class CacheBuilder {
     constructor();
     private get config();
     private key;
-    remember(key: string, callback: Function): Promise<any>;
+    remember(key: string, callback: Function, ttl?: number): Promise<any>;
     has(key: string): Promise<boolean>;
     get(key: string): Promise<any>;
-    add(key: string, value: any): Promise<boolean>;
-    put(key: string, value: any): Promise<boolean>;
+    add(key: string, value: any, ttl?: number): Promise<boolean>;
+    put(key: string, value: any, ttl?: number): Promise<boolean>;
     forget(key: string): Promise<void>;
+    increment(key: string): Promise<number>;
+    decrement(key: string): Promise<number>;
 }