+++ facades/Cache.js
@@ -1,7 +1,7 @@
 import CacheBuilder from "../builders/CacheBuilder";
 export default class Cache {
-    static async remember(key, callback) {
-        return new CacheBuilder().remember(key, callback);
+    static async remember(key, callback, ttl) {
+        return new CacheBuilder().remember(key, callback, ttl);
     }
     static async has(key) {
         return new CacheBuilder().has(key);
@@ -9,13 +9,19 @@ export default class Cache {
     static async get(key) {
         return new CacheBuilder().get(key);
     }
-    static async add(key, value) {
-        return new CacheBuilder().add(key, value);
+    static async add(key, value, ttl) {
+        return new CacheBuilder().add(key, value, ttl);
     }
-    static async put(key, value) {
-        return new CacheBuilder().put(key, value);
+    static async put(key, value, ttl) {
+        return new CacheBuilder().put(key, value, ttl);
     }
     static async forget(key) {
         return new CacheBuilder().forget(key);
+    }
+    static async increment(key) {
+        return new CacheBuilder().increment(key);
+    }
+    static async decrement(key) {
+        return new CacheBuilder().decrement(key);
     }
 }